address feedback

+++ go/lib/sciond/mock.go
@@ -97,7 +97,7 @@ func (m *MockConn) Paths(dst, src addr.IA, max uint16, f PathReqFlags) (*PathRep
 			PathReplyEntry{
 				Path: FwdPathMeta{
 					Interfaces: pathInterfaces,
-					ExpTime:    uint64(time.Now().Add(spath.MaxTTL * time.Second).Unix()),
+					ExpTime:    uint32(time.Now().Add(spath.MaxTTL * time.Second).Unix()),
 				},
 				HostInfo: HostInfo{
 				// TODO(scrye): leave nil for now since no tests use this
